e2e refactor wip.
