+++ e2e/nakama_client.go
