+++ e2e/tester/game/comp/location.go
