+++ e2e/tester/game/comp/player.go
