+++ e2e/tester/game/msg/join.go
