+++ go.work
@@ -4,10 +4,9 @@ use (
 	assert
 	cardinal
 	evm
-	internal/e2e
+	e2e
+	e2e/tester
 	internal/e2eutils
-	internal/e2e/tester/game
-	internal/e2ebenchmark/tester/game
 	internal/nakama
 	internal/nakamabenchmark
 	relay/nakama
